Move open_stream / read_stream / devices onto Endpoint; drop adb_channel
Three free functions in adb_channel.py already took an Endpoint as
their first argument and called endpoint.connect_to_adb() — they're
naturally methods on Endpoint. Promote them, then drop the now-tiny
adb_channel module entirely.

- endpoint.py: add Endpoint.open_stream / .read_stream / .devices
  (the latter renamed from list_adb_devices).
- adb_proxy.py: callers use endpoint.read_stream(...) /
  endpoint.devices(); AdbProxy.open_stream forwards to
  self.device_endpoint.open_stream. Move the one-line device_path
  helper here (its only caller).
- adb_channel.py: deleted.
